libosdp: Replace circular buffer command queue with linked list queue
The commands sent from application layer were queued to an adhoc
circular buffer and then popped by phy layers to serialize them. This
method was done as a stop-gap measure only but, managed to survive very
long. This commit replaces it with a linked list queue implementation.

This patch may introduce bugs. Also, it makes changes to osdp.h, so a
version bump is needed.

Signed-off-by: Siddharth Chandrasekaran <siddharth@embedjournal.com>