Fix sequence number 0 sent by CP does not reset connection in PD
As per the specification, when CP sends a sequence number 0, the PD must
reset all states it held and treat this message as the first message in
communication sequence. PD implementation was not handling this case
correctly where it was expecting each message to have incremented
sequence numbers.

This this by setting pd->seq_number to -1 so it is treated as the first
message since boot of PD.

Fixes: #1

Signed-off-by: Siddharth Chandrasekaran <siddharth@embedjournal.com>