Make PD message queue channel owner, responsible for cleanup
Both, CP and PD were cleaning up the message queue upon exit but only PD
is actually responsible for setting up the queue. This was causing issues
with repeated CP restarts as it removed the message queue the last time
the osdpctl process exited.

Now, we designate the PD as the owner of the message queue and make them
responsible for cleaning up those queues. The advantage of this method
is that we can start off some PDs and then forget about when working
with CP code.

Signed-off-by: Siddharth Chandrasekaran <siddharth@embedjournal.com>