Reset secure channel status of PD after a timeout
As per the specification, when there is a communication gap, long enough
to be treated as timeout condition, any pre-established secure channel
must be discarded.

So far, PD implementation wasn't discarding the SC status when a timeout
occurs while CP was doing this correctly. This wasn't too much of a
problem as the next retry from CP would be plain text and would force
the PD to loose it SC status too. Even so, a malicious CP could try to
continue operation in secure channel and potentially hijack the session.

Fix this issue by resetting secure channel status after a timeout
condition in PD.

Signed-off-by: Siddharth Chandrasekaran <siddharth@embedjournal.com>