libosdp: Resolve PD address / offset inconsistencies
Some places the PD is referred to with `pd_info_t *` array offsets. In
others (when pd replies) pd address is sent as pd identifier. For an
application, the PD must always be identifiable with one integer to
avoid confusion.

Fix this issue by using `pd_info_t *` array offset as the single point of
reference for a PD in CP mode. In PD mode, since there is only one PD,
no indication will be passed.

Fixes: #17

Signed-off-by: Siddharth Chandrasekaran <siddharth@embedjournal.com>