Fix osdp_event_cardread length bits/bytes mixup
Data is in bits if format is OSDP_CARD_FMT_RAW_* or bytes when format is
OSDP_CARD_FMT_ASCII. Update the comment to indicate this.

pyosdp was not expecting length field as it didn't expect bits in data.
Add support for this when format is set to OSDP_CARD_FMT_RAW_*

Fixes: #26

Signed-off-by: Siddharth Chandrasekaran <siddharth@embedjournal.com>