cp: Fix master keyset functionality
CP applications cannot set an arbitrary SCBK to connected PDs but only
choose to change its own master key. When a master key change is
requested, LibOSDP should prepare suitable SCBKs and send it to the
corresponding PDs which have an active SC.

Instead of exposing a new method to set master key, use existing struct
osdp_cmd_keyset by setting the key type to 0 (which is not defined in
the OSDP specification).

Related-To: #42

Signed-off-by: Siddharth Chandrasekaran <siddharth@embedjournal.com>