cp: Add new flag FLAG_SC_DISABLED when master_key == NULL
When CP is setup with master_key == NULL, there is no point in trying
to initiate a SC with the PDs. Add a flag FLAG_SC_DISABLED to stop such
attempts.

Note: If a PD was instructed to be setup with OSDP_FLAG_ENFORCE_SECURE,
then that PD will error into offline.

Signed-off-by: Siddharth Chandrasekaran <siddharth@embedjournal.com>