cp: Add bounded exponential back-off for offline retries
CP can go offline for a lot of reasons; some of them are transient (such
as when a PD misses the timing guarantees for a command because it is
busy doing other stuffs) while others are more long term (such as when
the PD is powered off). To handle both these cases in the most
efficiently, introduce a bounded, exponential, back-off for offline
retries.

That is, the CP would retry communication after 1s, 2s, 4s, 8s, ... till
it hits an upper bound (300s) and then on retry at that interval.

Closes: #48

Signed-off-by: Siddharth Chandrasekaran <siddharth@embedjournal.com>