Fix: Allow keyset with SCBK; guard master_key route with ENFORCE_SECURE
When the application want to start up in install mode and set a SCBK
directly (without deriving it from the master_key) cp_cmd_dispatcher()
should facilitate this based on PD_FLAG_HAS_SCBK. Until not this flow
was not permitted and libOSDP was using the master_key to generate the
SCBK (which was potentially all 0s).

Also, since the latest version of the specification recommends that we
don't use a master_key based derivation, guard it with our favorite
ENFORCE_SECURE flag.

Related-to: #60, #52
Signed-off-by: Siddharth Chandrasekaran <sidcha.dev@gmail.com>