cp: Allow applications to KEYSET with SCBK or master_key
Now applications can perform a key set operation with master_key or with
a dedicated SCBK without doing a compute_scbk().

Note: There is still one missing part of this puzzle: when a keyset with
a SCBK was performed, PD_FLAG_HAS_SCBK must be set on that PD so that
the SCBK survives secure channel restarts. As of now libOSDP does not
have a way to perform post-command-success actions. When we do have such
a feature, we should set this flag from that handler.

Creating a new issue #62 to track this gap.