libosdp: Cleanup error handling between phy and CP/PD
Phy layer defines a bunch of generic error codes that the CP/PD
implementation addapt according to their needs. PD uses this feature
extensively but CP was not catching all error codes from phy with a
generic CP error code which looks like a bug waiting to be reported.

While at it, fix the BUSY=2 error code hack and introduce a new phy
error code when PD phy wants to signal a NACK to the CP.

Signed-off-by: Siddharth Chandrasekaran <sidcha.dev@gmail.com>