libosdp: Remove __weak directive on crypto wrapper methods
It appears empty __weak methods are NULL-ed (instruction pointer is set
to 0) out in GCC for some reason and we segfault; as usual clang works
just fine.

Now that crypto library wrappers are moved into their own directory, the
need for having them as weak refererence is not pressing anymore. If
someone wants to add a new library wrapper, they can create a new file
under src/crypto/ and send a PR for it :)

Signed-off-by: Siddharth Chandrasekaran <sidcha@amazon.de>
Cc-Team: kaos-brimstone <kaos-brimstone@amazon.com>
Signed-off-by: Siddharth Chandrasekaran <sidcha.dev@gmail.com>