libosdp: Add command/reply data buffer tracing option
CONFIG_PACKET_TRACE has been around since epoch and served it's purpose
very well. The drawback of this method of tracing is that is is
completely useless when secure channel is enabled and there is a command
level bug we are debugging. Add new configure.sh and CMake options
CONFIG_OSDP_DATA_TRACE (--data-trace) to print the command/reply buffer
as they are sent/received.

Signed-off-by: Siddharth Chandrasekaran <sidcha.dev@gmail.com>