pytest: Add pytest for more sophisticated tests
The existing poor man's test suite for libosdp is not powerful enough to
test the kind of things that we ought to be testing at this stage of
maturity of the project. This was made painfully obvious when trying to
debug the SCBK clobber issue #60.

Add pytest integration and some initial framework to start writing more
complicated tests.

Signed-off-by: Siddharth Chandrasekaran <sidcha.dev@gmail.com>