libosdp: Double SC timeout to be a bit more flexible
Value of OSDP_PD_SC_TIMEOUT_MS is not defined by OSDP but is a LibOSDP
set limit. It was derived as 2 times the value of OSDP_RESP_TOUT_MS
which is defined by OSDP to be 200ms (section 5.7 Timing). The spec also
mentions that the PD should respond within 3ms typically and if it
cannot, it should send REPLY_BUSY.

Although this allows a CP enough time to poll to a subsequent PD it
seems a bit aggressive so lets double it for the time being.

Related-to: #68

Signed-off-by: Siddharth Chandrasekaran <sidcha.dev@gmail.com>