pytest: Dont abuse pyosdp namespace
Currently the python bindings for LibOSDP uses the "pyosdp" namespace.
pytests has its own wrapper around the python C extension to simplify
some testing access which also used "pyosdp" name this module. This is
causing some level of confusions occasionally.

Rename it as "testlib" to be more descriptive of what is it.

Signed-off-by: Siddharth Chandrasekaran <sidcha.dev@gmail.com>