osdp.h: Add comments to clarify osdp_MFG structure
OSDP specification v2.2 makes the structure of osdp_MFGREP to be fixed
at 4 bytes (instead of the variable length before). This allowed some
flexibility to defined some minimal structure to these commands so the
whole length prefixed data thing was a LibOSDP specific extension (IMO,
this is how they should have defined the struct to begin with).

The good thing is that when the length field is set to zero, the current
code behaves exactly as the new spec says it should so no harm done.
This commit adds a comment to clarify this.

Related-to: #75
Signed-off-by: Siddharth Chandrasekaran <sidcha.dev@gmail.com>