libosdp: Add a softname for PDs in the info struct
During logging, when there are more than one PDs connected to a CP, the
messages appear in the same log stream and hence it is essential to
distinguish the PD by a name. Till now, LibOSDP was setting this name.
Since that is not as flexible as we'd like it to be, move this into the
pd_info_t struct so the user can provide a name of their choosing.

This patch enabled this variable all over the place: pyosdp, testlib,
and osdpctl.

Signed-off-by: Siddharth Chandrasekaran <sidcha.dev@gmail.com>