libosdp: Add a ring buffer to store the RX data
The current flat buffer we have to store the RX data though having the
advantage of being able to pass to osdp_sc.c directly for decrypting and
encrpyting, actually imposes a lot restrictions and comes with a
performance penality. For starters, we have  a memmove operation that
was a ductape solution for not having a ring buffer and pulling data
from it at the time.

On top of that, gh-issue: #84 reqires us to drop an arbitary number of
leading MARK bytes. As a perparatory step, this patch introduces a ring
buffer for RX data and pulls exactly _one_ packet into a working buffer
and operates on it.

Related-to: #84
Signed-off-by: Siddharth Chandrasekaran <sidcha.dev@gmail.com>