libosdp: Make osdp_cmd_mfg and osdp_mfgrep identical
As pointed out by adakus@ in zephyr review, we may want to allow PDs to
send unsolicited MFGREP responses in future. To allow for that without
having to deal with it as a bug, let us fix this one proactively.

This commit also fixes a typo and bumps utils submodule to latest
version.

Link: https://github.com/zephyrproject-rtos/zephyr/pull/46449
Signed-off-by: Siddharth Chandrasekaran <sidcha.dev@gmail.com>