pyosdp, pytests: Add test case for mfgrep event
osdp_MFGREP event notification is was reported to be broken in #96. This
commit adds a test to add coverage for this case and ensures that it is
indeed the case.

Also, since this event is not enabled in the python module, it could not
be used in the pytest suite; this patch enables those parts too.

Related-to: #96
Signed-off-by: Siddharth Chandrasekaran <sidcha.dev@gmail.com>