unit-test: Fix a use-after-free bug in test async runner
Unit test async runner was populating complete function which was
freeing the work pointer. The work cancel requester looks up the work
pointer after it has been free-d to check if the work status has changed
to "completed". Fix this by not populating the complete_fn and freeing
the work struct manually.

Signed-off-by: Siddharth Chandrasekaran <sidcha.dev@gmail.com>