libosdp: file: Promote to be a fist-class citizen
The file API was developed as a plugin to libosdp to ensure it doesn't
clutter the rest of the library while its design was still being
formalized. Since then, it has stabilized quite a bit and has some
testing creds too. This patch promotes it to be a fist-class citizen in
LibOSDP (no longer can be compliled out).

Signed-off-by: Siddharth Chandrasekaran <sidcha.dev@gmail.com>