file: Fix bug in return value processing in build_command
When build command fails, the entire command has to be failed. Instead,
we were merely breaking causing a zero length command to be sent over to
the PD. This could be one potential candidate that could cause an issue
like #100 but since it was wasn't reproduced/tested, we cannot say for
sure.

Related-to: #100
Signed-off-by: Siddharth Chandrasekaran <sidcha.dev@gmail.com>