libosdp: Limit max packet size to reported peer RX size
Both CP and PD have means of reporting their RX buffer size to the other
(CP by a command; PD by the capability report) and both must fit the
message bytes to the peer's limit.

Till now, LibOSDP was collecting this info but wasn't actually limiting
the packet size with it. This patch adds support for that.

Fixes: #102
Signed-off-by: Siddharth Chandrasekaran <sidcha.dev@gmail.com>