phy: Allow sequence repeat packets
OSDP specification supports command retires by allowing the CP to resend
the same command without incrementing the sequence number which prompts
to PD to resend the last response. One major use case for this feature
is to work around RS485 channel errors where the CP find the received
packet to have checksum errors.

Handling this feature in the way the specification intended, makes the
code look ugly, complicates things or wastes resources. Instead, we will
just ignore such non-movement in packet sequence and re-process the
packet all over again as it it was the first time we are seeing it.
Since this is an unhappy case, keeping it simple is more important than
performance.

Fixes: #41
Signed-off-by: Siddharth Chandrasekaran <sidcha.dev@gmail.com>