libosdp: Add support for tamper and power status events
Tamper and power status are owned by PD and reported to CP as an event.
This patch adds support for PD app to send a status event and a CP to
receive this event and process it.

While at it, also deprecate the internal tamper and power status
tracking as we have no use for it.

Fixes: #105
Signed-off-by: Siddharth Chandrasekaran <sidcha.dev@gmail.com>