libosdp: Remove command_complete_callback API
command_complete_callback is exposing low level OSDP command completion
to the application layer. This is going against rules of abstraction
that we'd like to have. What could be useful is a callback for osdp.h
exposed commands and events but this callback as it is is not adding
much value.

Let's remove this callback for now and then think about how to enhance it
later.

Signed-off-by: Siddharth Chandrasekaran <sidcha.dev@gmail.com>