libosdp: Add new API to set a log callback function
Until now, LibOSDP was doing all the log message formatting only allowed
the users to provide a puts() like function that dispersed the log
message to some medium. This API is limiting to some users who have
their own logger configured in the application and like to pass OSDP log
messages into the same stream.

To handle such cases, add a new API that allows a user to set a log
callback function that gets called with the raw log message and other
meta data about the log message such as log level.

Fixes: #139
Signed-off-by: Siddharth Chandrasekaran <sidcha.dev@gmail.com>