libosdp: cp: Refactor API command handling (translating) logic
When we receive a command from the user, the API method was doing the
initial validation and also translating those commands into internal
osdp commands. Although this works, it doesn't allow for some level of
flexibility of internal methods. Refactor this logic into a new function
called cp_translate_cmd() which is invoked when the command is dequeued
from the FSM.

Signed-off-by: Siddharth Chandrasekaran <sidcha.dev@gmail.com>