libosdp: cp: Add different query options for OSDP_CMD_STATUS
Till now, OSDP_CMD_STATUS could only query for local status (power and
tamper) but there are two other options to ask for from the PD: input
and output.

This commit add an inner struct for OSDP_CMD_STATUS which allows users
to specify which type of status they want to query.

Signed-off-by: Siddharth Chandrasekaran <sidcha.dev@gmail.com>