rust: Split bindgen generated bindings to a separate -sys crate
Cargo has very rigid rules on how a package can be distributed. Since
this conforms with the wider rust community' expectations, we will
adhere to this standard and release a separate -sys crate and consume it
here. It does add a bit of hassle for us with the back and forth.. only
time will tell of this split serves other purposes than just to appease
cargo.

Signed-off-by: Siddharth Chandrasekaran <sidcha.dev@gmail.com>