libosdp: phy: Ignore reply with PD address.MSB cleared
According to the specification, all response packets must have the
address byte with the MSB set. Till now, the CP treated this as a fatal
error, causing the secure channel to be lost. Since it does not harm it
in any way, we could just safely ignore that packet and keep proceeding.

Signed-off-by: Siddharth Chandrasekaran <sidcha.dev@gmail.com>