libosdp: phy: Fix packet trace mark byte logging
When packet trace is enabled, we have an issue where logs on one side
print the mark byte (0xff) and the other side ignores it. Although this
does not cause any issues, it causes the question if the mark byte was
sent at all.

Fix this issue by inserting the mark byte into the packet_buf manually
if we have parsed it. Also, this makes all methods uniformly include
mark byte handling so as to make it's presence obvious.

Signed-off-by: Siddharth Chandrasekaran <sidcha.dev@gmail.com>