libosdp: cp: Rework CP state machine
The CP state machine that exists today is a partial fixed up version of
its predecessor. It was never cleanly rewritten to account for the
changes in other parts of the library.

This patch completely rewrites the CP state machine fixing may cases
where we had unnecessary states which made it harder to follow the
control flow.

Signed-off-by: Siddharth Chandrasekaran <sidcha.dev@gmail.com>