libosdp: Reduce command/event pool size
Since we bumped the size of event data len to 128 in #152, the
command/event pool size of 32 uses almost 4KiB of RAM. To fix this CP
state machine was completely rewritten to not use the command queue
entirely. With this change, the command/event queue is exclusively meant
for commands and events sent from the application to libosdp and does
not need to be as large as 32.

Reduce this queue size to 4 -- thereby reducing the total memory used to
just 512 bytes.

Signed-off-by: Siddharth Chandrasekaran <sidcha.dev@gmail.com>