libosdp: Rework status command and event handling
It was reported in #154 that LibOSDP has no way for PD app to answer
some commands. To fix this, a bit of rework was needed in how we look at
the status event and command.

Note: This commit breaks API and ABI since we have changed command and
event numbers (amd removed existing ones). Since we are in the works for
a major relase from master, and we have a stable 2.4.x branch, this
should be okay.

Fixes: #154
Signed-off-by: Siddharth Chandrasekaran <sidcha.dev@gmail.com>