libosdp: cp: Fix broken keyset for an install_mode PD
When the PD is in install mode, the CP should be able to establish a
secure channel with the default SCBK and then use that sesssion to set a
new key (the one the CP recieved for that PD during setup).

This used to work before but was broken during the CP state machine
cleanup done in b4cfb576ab and missing tests for this case allowed that
regression to go unnoticed (this issue was addressed in the last
commit).

This commit fixed the regression.

Fixes: #157
Fixes: b4cfb576ab603fdd434a1883d475ac00ab84e61c
Signed-off-by: Siddharth Chandrasekaran <sidcha.dev@gmail.com>