python: Remove code referring to CMakeLists.txt version
Python build was reading and parsing CMakeLists.txt from the parent dir
to get the project name and version. Although this reduces one point
that can fall out of sync during a release, it caused python builds to
break when built/installed from pip.

This commit fixes this issue by hard coding the version in python/setup.py
and then extends the release tooling to also update this file.

Related-to: #160
Signed-off-by: Marco Nilsson <denravonska@gmail.com>
Signed-off-by: Siddharth Chandrasekaran <sidcha.dev@gmail.com>