python: Remove concrete "channel" implementations
LibOSDP creates a notion of a "channel" something which can be written
to and read from. This abstraction allows each implementer to have
whatever transport method they like to use to talk with the CP/PD and
then just apply the library on top of it.

The cpython layer started creating concrete types for channels and then
passing along a channel_type config var (which can be unix, fifo, serial
etc.,) which is used to setup a channel internally using the c-utils
library which also exports a module called channel to solve this
specific problem.

While this approach makes the life of the python user simple, it
introduced the classic dependency inversion problem: a high level module
(cpython) is depending on concrete types exposed by a low level module
(c-utils). This issue prevents LibOSDP from running on platforms such as
Windows where it can technically run.

This patch addresses this issue by removing all concrete types from
cpython and asking all the implementers to implement the Channel
abstract class and then pass it in PdInfo during CP/PD creation. While
this make the python code a little bit more verbose, it is the right and
most flexible thing to do.

This commit also fixes a minor bug in that the Python layer requested
log_level was being ignored and the default value LOG_INFO was applied.

Signed-off-by: Siddharth Chandrasekaran <sidcha.dev@gmail.com>