libosdp: Fix broken SC due to missing sc_init call
When the CP state machine was reworked, the call to sc_init was missed
before starting a secure channel. This caused the CP random data to be
zeros leading to SC having no effect at all. To fix this issue, a little
bit of code rearrangement was needed.

As a side-effect, we no longer need to change call sc_keyset_complete()
from phy layer which was a known layer violation. This make the control
flow fit much better with the rest of the new CP design.

Signed-off-by: Siddharth Chandrasekaran <sidcha.dev@gmail.com>