Fix bug in SC setup due to missing call to cp_keyset_complete
When SC is setup for the first time when the PD is in install mode, the
CP state machine should call cp_keyset_complete to clear the current SC
session so a new one can be setup.

During refactoring, this was broken and hence there was a failure during
initial SC handshake.

Signed-off-by: Siddharth Chandrasekaran <sidcha.dev@gmail.com>