phy: Allow multiple mark bytes before SOM
The specification requires the transmitting devices to allow an
equivalent of 2 character time as idle time between packets. It also
mentions that it must drive the line to all-high (done by sending 0xFF
over the wire) for a minimum of 1 character width.

The combination of these two should yield a pause of 2 char, and a mark
of 1 char (UART start bit is 0). But some devices my chose to send more
than one mark byte and we will be a bit more lenient to allow this.

Signed-off-by: Siddharth Chandrasekaran <sidcha.dev@gmail.com>