pytest: Disable status tests
Due to the chagnes to which allows it to retain a PD online for 8
seconds (after that the CP takes 300 seconds to come back online) this
test does not make sense anymore.

In future, this test must be migrated to unit-tests so we can kick out a
PD from offline state with some off-band signal so the tests can be
completed faster.

Signed-off-by: Siddharth Chandrasekaran <sidcha.dev@gmail.com>