python: Enable packet trace builds by default
The python module is mostly used for debugging and testing LibOSDP.
Anyone using it in production is probably using the C library. So it
makes sense to enable all debugging and testing features by default on
the python builds so they can be pip installed to test things out
quickly.

Signed-off-by: Siddharth Chandrasekaran <sidcha.dev@gmail.com>