file: Fix multiple assumptions about target endianess
When file ops was developed as PoC, it was simpler to rely of the fact
that x86 is little endian and just typecast the buf into target data
structures. This may not be the case all the time. This patch gets rid
of those typecasts and replaces them with hand packing code.

Signed-off-by: Siddharth Chandrasekaran <sidcha.dev@gmail.com>