cp: Fix typo in flag CP_REQ_OFFLINE value
When adding CP_REQ_OFFLINE flag, the value was set to 0x01 which was
previously assigned to CP_REQ_SC_RESTART. This caused the PD to restart
secure channel instead of going offline.

The second issue is that the restart secure channel path is not guarded
by a suitable check to see if this PD can do secure channel at all.
Although we should never reach this code path in the normal flow, it's
nice to know that these code paths converge at some common point where
we check for the PD's SC capability and then proceed to restart SC. A
later patch will address this issue.

Related-to: #191
Fixes: 4225650a4327b22ef3e5841b68a84207987346e2
Signed-off-by: Siddharth Chandrasekaran <sidcha.dev@gmail.com>