python: Fix LED commands with temporary and permanent records
Some times, the CP sends a permanent LED command and also fills the
temporary command structure to cancel any temporary LED activities.

The python implementation has a limitation (undocumented) that it can
pass on either temporary or permanent records only causing issue for a
user (see #200).

This patch fixes this issue by adding a new member "cancel_temporary"
which conveys this intent to the application.

Fixes: #200
Signed-off-by: Siddharth Chandrasekaran <sidcha.dev@gmail.com>