Fix handling of empty SCS17/SCS18 messages
Add a flag to allow for non-compliant messages for SCS17/SCS18 messages.  Some
OSDP implementations will send an SCS17/SCS18 message with no data when in
secure mode.  This is not correct according to the standard.  If this message
is received with libosdp, the current behavior is to drop the message and
print an error.

This patch adds a new flag, FLAG_ALLOW_EMPTY_ENCRYPTED_DATA_BLOCK to allow for
non-standard behavior of an empty data block with SCS17/SCS18.  This is only
when libosdp is used as a PD.