pd: Drop check for allowed baud rates in COMSET
In the past, the specification used to state explicit baud rates that the
PD can operate in (last, that's how I remember it). Now it just says at
least one of those bauds need to be supported.

This means, LibOSDP is not in the right place to satisfy this
requirement correctly. So we will just transfer this responsibility to
the implementer.

Fixes: #212
Signed-off-by: Siddharth Chandrasekaran <sidcha.dev@gmail.com>