API: PD: Add new COMESET_DONE command
When COMSET command succeeds, the PD is supposed to respond back to the
CP with the current setting and then switch to the new setting. LibOSDP
does not have a way to notify the PD app (after the REPLY_COM response)
to switch to the new channel.

This patch a adds a new command OSDP_CMD_COMSET_DONE which is the same
as OSDP_CMD_COMSET but only that it is a cue for the PD app to "commit"
the changes.

Related-to: #215
Co-Developed-by: Siddharth Chandrasekaran <sidcha.dev@gmail.com>
Signed-off-by: Samuel Gouraud <sgouraud@fdi-access.com>