pd: Add missing call to osdp_sc_init()
Commit 8200b951f6956c523e1c8107e9368476939310f2 unwittingly removed a
call to osdp_sc_setup() from pd_decode_command(). This caused the PD to
setup secure channel with client_uid set to zeros.

Apart from that, this also causes the issue with crypt subsystem being
uninitalized in PD mode. This doesn't affect the embedded TinyAES module
as it down not have a init phase.

Signed-off-by: Siddharth Chandrasekaran <sidcha.dev@gmail.com>