API: provide export macros locally
The export macros generated by cmake does not handle the windows build
case correctly. Also, it only works for the cmake build and everything
else was defining those macros to be no-ops.

To Fix this situation, let's locally provide those macros. Also, change
the default visibility of symbols to hidden on all platforms.

Signed-off-by: Siddharth Chandrasekaran <sidcha.dev@gmail.com>