pd: Fix bug in MFGREP events from application
MFGREP was assumed to be sent only in response to a MFG command. But it
turns out they can be sent as a POLL response too. This means, we handle
externally arriving MFGREP commands too.

This patch fixes this issue by allowing the app to send an unsolicited
MFGREP event in addition to the existing method of responding to the MFG
command.

Fixes: #240
Signed-off-by: Siddharth Chandrasekaran <sidcha.dev@gmail.com>