PD: Fix non-compliant REPLY_RMAC_I status byte
The PD implementation incorrectly sets the secure channel status byte in
SCS_14 (REPLY_RMAC_I) response to 0 when CP cryptogram verification
fails instead of 0xff.

According to OSDP specification section D.1.4.5, Sec_blk_data[0] should
be set to 1 (success) or 0xff (failure) to indicate status. But it does
not make it clear if the RMAC_I should be sent when the verification
fails. It just "feels" wrong to return a value (RMAC_I) on an error
path.

Since the specification also says the PD can respond with a NAK(5) when
the verification fails, that option is unambiguous, let's resort sending
the RMAC_I on success and NAC(5) on failures.

Fixes: #245
Signed-off-by: Siddharth Chandrasekaran <sidcha.dev@gmail.com>