CP: Fix bug in sequence number progression
Commit 7f6d3c358 introduced a fix for the long-standing sequence number
mismatch issue by an explicit sequence progression logic. By this logic,
when there are no errors, the sequence number is progressed. But it
failed to handle the case when the PD respond with a valid packet, and
the CP finds an application layer error in the packet, the sequence
number must be progressed as the PD would have (correctly) incremented
it's sequence number.

To fix this, let's introduce a new CP error code to use for any future
application layer errors and use that to hook into the sequence
progression logic.

Fixes: #244
Fixes: 7f6d3c358efc741f02de143782fde9e86d7abc35
Signed-off-by: Siddharth Chandrasekaran <sidcha.dev@gmail.com>