cp: Remove unused flag PD_FLAG_HAS_SCBK
PD_FLAG_HAS_SCBK was used to decide if the SCBK must be preserved when
secure channel is initialized. This was from an era where we had master
key based key derivation where the SCBK was derived.

Since master-key has long been depreciated and subsequently removed, we
can get rid of that flag too.

Signed-off-by: Siddharth Chandrasekaran <sidcha.dev@gmail.com>